pyproject.toml
[project]
name = "agentic-eval"
version = "0.1.0"
description = "Patent Idea Relevance Scoring Engine - Agentic Pipeline"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai-agents>=0.6.4",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
]
